OGS-3: Storage initialization restructuring implementation complete. To be tested.

+++ net.jadoth.storage/src/net/jadoth/storage/types/StorageDataFile.java
@@ -174,7 +174,7 @@ public interface StorageDataFile<I extends StorageEntityCacheItem<I>> extends St
 			(this.prev.next = this.next).prev = this.prev;
 		}
 
-		final void registerGap(final int length)
+		final void registerGap(final long length)
 		{
 			this.fileTotalLength += length;
 		}
